+++ pycobertura/reporters.py
@@ -19,31 +19,29 @@ headers_without_missing = ["Filename", "Stmts", "Miss", "Cover"]
 
 
 class Reporter:
-    def __init__(self, cobertura):
+    def __init__(self, cobertura, ignore_regex=None):
         self.cobertura = cobertura
+        self.ignore_regex = ignore_regex
 
     @staticmethod
     def format_line_rate(line_rate):
         return f"{line_rate:.2%}"
 
     def get_report_lines(self):
+        filenames = self.cobertura.files(ignore_regex=self.ignore_regex)
         lines = {
-            "Filename": self.cobertura.files().copy(),
+            "Filename": filenames.copy(),
             "Stmts": [
-                self.cobertura.total_statements(filename)
-                for filename in self.cobertura.files()
+                self.cobertura.total_statements(filename) for filename in filenames
             ],
-            "Miss": [
-                self.cobertura.total_misses(filename)
-                for filename in self.cobertura.files()
-            ],
+            "Miss": [self.cobertura.total_misses(filename) for filename in filenames],
             "Cover": [
                 self.format_line_rate(self.cobertura.line_rate(filename))
-                for filename in self.cobertura.files()
+                for filename in filenames
             ],
             "Missing": [
                 stringify(self.cobertura.missed_lines(filename))
-                for filename in self.cobertura.files()
+                for filename in filenames
             ],
         }
         lines["Filename"].append("TOTAL")
@@ -171,10 +169,19 @@ class HtmlReporter(Reporter):
 
 
 class DeltaReporter:
-    def __init__(self, cobertura1, cobertura2, show_source=True, *args, **kwargs):
+    def __init__(
+        self,
+        cobertura1,
+        cobertura2,
+        ignore_regex=None,
+        show_source=True,
+        *args,
+        **kwargs,
+    ):
         self.differ = CoberturaDiff(cobertura1, cobertura2)
         self.show_source = show_source
         self.color = kwargs.pop("color", False)
+        self.ignore_regex = ignore_regex
 
     def format_line_rate(self, line_rate):
         return f"{line_rate:+.2%}" if line_rate else "+100.00%"
@@ -213,17 +220,17 @@ class DeltaReporter:
         return self.color_number(f"{total_misses:+d}") if total_misses else "0"
 
     def get_report_lines(self):
+        filenames = self.differ.files(ignore_regex=self.ignore_regex)
         diff_total_stmts = [
-            self.differ.diff_total_statements(filename)
-            for filename in self.differ.files()
+            self.differ.diff_total_statements(filename) for filename in filenames
         ]
 
         diff_total_miss = [
-            self.differ.diff_total_misses(filename) for filename in self.differ.files()
+            self.differ.diff_total_misses(filename) for filename in filenames
         ]
 
         diff_total_cover = [
-            self.differ.diff_line_rate(filename) for filename in self.differ.files()
+            self.differ.diff_line_rate(filename) for filename in filenames
         ]
 
         indexes_of_files_with_changes = [